# src/main/resources/static/openapi/openapi.yml
openapi: 3.0.0
info:
  title: Chat API
  version: 1.0.0
  description: FeitApi for university of UNIZA, FEIT
servers:
  - url: http://localhost:8080/
paths:
  /api/v1/blog/tags:
    get:
      summary: Find blog posts by tags
      operationId: getPostsByTags
      tags:
        - blog
      parameters:
        - name: tagNames
          in: query
          description: List of tag names to filter posts by
          required: true
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '400':
          description: Invalid request
  /api/v1/tags:
    get:
      summary: Get all tags
      operationId: getAllTags
      tags:
        - tags
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      summary: Create new tag
      operationId: createTag
      tags:
        - tags
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagRequest'
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid input
  /api/v1/tags/{id}:
    get:
      summary: Get tag by ID
      operationId: getTagById
      tags:
        - tags
      parameters:
        - name: id
          in: path
          description: ID of the tag
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found
    delete:
      summary: Delete tag by ID
      operationId: deleteTagById
      tags:
        - tags
      parameters:
        - name: id
          in: path
          description: ID of the tag
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Tag deleted successfully
        '404':
          description: Tag not found
  /api/v1/image:
    get:
      summary: pageable image list
      operationId: getAllImages
      tags:
        - image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                page:
                  type: integer
                  format: int64
                  description: Page number (zero-based)
                  default: 0
                size:
                  type: integer
                  format: int64
                  description: Number of items per page
                  default: 10
      responses:
        '200':
          description: Image list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageListPost'
    post:
      summary: Upload image
      operationId: uploadImage
      tags:
        - image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: URL of the uploaded image
  /api/v1/image/{id}:
    get:
      summary: get image by id
      operationId: getImage
      tags:
        - image
      parameters:
        - name: id
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Image found
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: URL of the image
        '404':
          description: Image not found

    delete:
      summary: delete image by id
      operationId: deleteImage
      tags:
        - image
      parameters:
        - name: id
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Image deleted successfully
        '404':
          description: Image not found
  /api/v1/blog:
    get:
      summary: Get all blog posts with pagination
      operationId: getAllPosts
      tags:
        - blog
      parameters:
        - name: page
          in: query
          description: Page number (zero-based)
          required: false
          schema:
            type: integer
            format: int64
            default: 0
        - name: size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int64

            default: 10
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageListPost'
    post:
      summary: Create a new blog post
      operationId: createPost
      tags:
        - blog
      requestBody:
        description: Blog post to create
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Title of the blog post
                content:
                  type: string
                  description: Content of the blog post
                tags:
                  type: array
                  items:
                    type: string
                  description: List of tags for the blog post
                mainImage:
                  type: string
                  format: binary
                  description: Main image for the blog post (base64 encoded)
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
  /api/v1/blog/{id}:
    get:
      summary: Get blog post by ID
      operationId: getPostById
      tags:
        - blog
      parameters:
        - name: id
          in: path
          description: ID of the blog post
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found

    put:
      summary: Update a blog post
      operationId: updatePost
      tags:
        - blog
      parameters:
        - name: id
          in: path
          description: ID of the blog post to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated blog post
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostRequest'
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
        '404':
          description: Post not found

    delete:
      summary: Delete a blog post
      operationId: deletePost
      tags:
        - blog
      parameters:
        - name: id
          in: path
          description: ID of the blog post to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Post deleted successfully
        '404':
          description: Post not found
...
components:
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the post
        title:
          type: string
          description: Title of the blog post
        content:
          type: string
          description: Content of the blog post
        author:
          type: string
          description: Author of the blog post
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        tags:
          type: array
          items:
            type: string
          description: List of tags for the blog post
        mainImage:
          type: string
          format: byte
          description: Main image for the blog post (base64 encoded)
      required:
        - title
        - content
    PostRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the blog post
        content:
          type: string
          description: Content of the blog post
        tags:
          type: array
          items:
            type: string
          description: List of tags for the blog post
      required:
        - title
        - content
    PageListPost:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Post'
          description: List of posts for the current page
        totalElements:
          type: integer
          format: int64
          description: Total number of posts
        nextPage:
          type: string
          nullable: true
          description: URL to the next page if available
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the tag
        name:
          type: string
          description: Name of the tag
      required:
        - name
    TagRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the tag
      required:
        - name